Updated Tinybird tracker endpoint to use versioned API endpoint (#17)
ref https://linear.app/ghost/issue/PROD-2270/update-the-ghost-docker-repo-to-use-the-apiv1page-hit-endpoint

The Analytics Service is now using a new versioned endpoint at `/api/v1/page_hit` instead of the `/tb/web_analytics` endpoint. This update's Ghost's configuration to use this new endpoint for the tracker requests.

Note: this will require a docker compose pull to get the latest version of the analytics service, which includes the new versioned endpoint.